The rendering of the block work area indicator has been modified. 修改了方块工作范围指示框的渲染 (#4767)
* refactor(client): 优化多部分预览不透明度及影响范围渲染逻辑

- 将配置中的多部分预览不透明度类型由float改为double，增强精度
- 在HudTooltipManager中新增解析作用范围方法，提升代码复用性
- 修改多处多部分预览相关代码，确保不透明度转换为float后再使用
- RenderEventListener新增影响范围持久化缓存，避免频繁计算和渲染
- 在渲染事件中根据游戏时间管理作用范围的显示时长，实现边框持续显示效果
- 优化智能放置器和结构扫描器的范围渲染，改用统一的颜色和持久化逻辑
- 精简渲染流程，添加null安全检查，提升稳定性和性能

* fix(client): 修复渲染时可能的空指针异常

- 引入 @Nullable 注解标记 currentPos 变量
- 在渲染缺失放大器轮廓和玻璃时增加 level 非空检查
- 防止 level 为 null 导致的渲染调用崩溃
- 提升渲染代码的健壮性和稳定性

* refactor(tooltip): 移除renderAffectRange方法及相关导入

- 删除了HudTooltipManager中的renderAffectRange方法实现
- 移除了不再使用的renderOutline静态导入
- 简化了代码，减少冗余方法和依赖
- 保持代码清洁，提升可维护性

* feat(tooltip): 更新奇点水晶物品提示文本

- 修改奇点水晶的普通工具提示为“由微型黑洞和微型白洞形成的水晶”
- 添加奇点水晶的按Shift显示的详细提示，说明其存储天体数据及与超维终端交互方法
- 同步更新en_us和en_ud语言文件中奇点水晶的提示文本内容

* refactor(client): 优化渲染事件监听器的范围边框逻辑

- 在 RangeOutline 记录中添加 BlockState，避免加载过时边框
- 处理 Level 为空时清理渲染状态，防止残留边框
- 修改方法由直接渲染改为登记范围，推迟渲染到统一处理
- renderPersistedOutlines 增加对 BlockState 变化检测，及时移除失效边框
- 精简并统一范围边框的添加流程，提升代码可维护性
- 修正并更新部分注释，增强代码可读性

---------

+++ src/main/java/dev/dubhe/anvilcraft/config/AnvilCraftClientConfig.java
@@ -109,11 +109,11 @@ public class AnvilCraftClientConfig {
 
     @Comment("Opacity of the ghost (solid) preview when placing multipart blocks")
     @BoundedDiscrete(min = 0.0, max = 1.0)
-    public float multiPartPreviewGhostOpacity = 0.3f;
+    public double multiPartPreviewGhostOpacity = 0.3;
 
     @Comment("Opacity of the outline preview when placing multipart blocks")
     @BoundedDiscrete(min = 0.0, max = 1.0)
-    public float multiPartPreviewOutlineOpacity = 0.5f;
+    public double multiPartPreviewOutlineOpacity = 0.5;
 
     @SerializedName("Ionocraft Backpack HUD")
     @CollapsibleObject
